refactor: split definitions constants and Instruction into modules
