+++ svm/accountant/crates/definitions/src/constants/seeds.rs
@@ -1,0 +1,33 @@
+//! PDA seed prefixes for every account type.
+
+/// PDA seed prefix for [`crate::PendingObservationsLayout`]. Full tuple:
+/// `(b"pending", chain_be, emitter, sequence_be, digest)`. The digest suffix
+/// lets fork/reorg observations accumulate in sibling buckets and binds each
+/// bucket to its digest, so no runtime digest-equality check is needed.
+pub const PENDING_OBSERVATIONS_SEED_PREFIX: &[u8] = b"pending";
+
+/// PDA seed prefix for [`crate::BalanceAccountLayout`]. Full tuple:
+/// `(b"account", chain_be, token_chain_be, token_address)`. Big-endian chain
+/// fields match the VAA wire format and the other seed derivations.
+pub const ACCOUNT_SEED_PREFIX: &[u8] = b"account";
+
+/// PDA seed prefix for [`crate::ChainRegistrationLayout`]. Full tuple:
+/// `(b"chain_registration", chain_be)`.
+pub const CHAIN_REGISTRATION_SEED_PREFIX: &[u8] = b"chain_registration";
+
+/// PDA seed prefix for [`crate::ModificationLayout`]. Full tuple:
+/// `(b"modification", sequence_be)`. Existence of this PDA enforces replay
+/// protection on the governance path.
+pub const MODIFICATION_SEED_PREFIX: &[u8] = b"modification";
+
+/// PDA seed prefix for the global-accountant authority that signs all
+/// `solana-noreplay` CPIs. Full tuple: `[b"noreplay_authority"]`. One global
+/// authority suffices because the noreplay namespace (`chain_be ‖ emitter`)
+/// already segregates per-emitter sequence spaces.
+pub const NOREPLAY_AUTHORITY_SEED_PREFIX: &[u8] = b"noreplay_authority";
+
+/// PDA seed prefix for Core Bridge's GuardianSet accounts. Full tuple:
+/// `(b"GuardianSet", guardian_set_index_be)`. Owned by `CORE_BRIDGE_PROGRAM_ID`;
+/// `submit_observations` re-derives this address to authenticate the GuardianSet
+/// account before reading guardian keys from it.
+pub const GUARDIAN_SET_SEED: &[u8] = b"GuardianSet";
